requirements.txt
django~=3.1.7
social-auth-app-django~=4.0.0
python-jose~=3.2.0
python-dotenv~=0.15.0
pytest-django~=4.1.0
black==20.8b1
psycopg2-binary==2.8.6
dj-database-url==0.5.0
whitenoise==5.2.0
gunicorn==20.0.4
sentry-sdk==0.20.3
github-contents==0.2
